@@ -1,12 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <PreserveCompilationContext>true</PreserveCompilationContext>
     <AssemblyName>dexih.api</AssemblyName>
     <OutputType>Exe</OutputType>
     <PackageId>dexih.api</PackageId>
     <UserSecretsId>aspnet-dexih.api-20161103030146</UserSecretsId>
-    <RuntimeIdentifiers>win-x64;linux-x64,osx-x64</RuntimeIdentifiers>
     <VersionPrefix>0.4.0-beta</VersionPrefix>
     <VersionSuffix>00001</VersionSuffix>
     <AssemblyVersion>0.0.1.1</AssemblyVersion>
@@ -14,7 +13,7 @@
     <Company>Data Experts Group</Company>
     <Product>Information Hub</Product>
     <DockerComposeProjectPath>..\..\docker-compose.dcproj</DockerComposeProjectPath>
-    <LangVersion>7.3</LangVersion>
+    <LangVersion>8</LangVersion>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\Dexih.App.Remote\src\dexih.remote.operations\dexih.remote.operations.csproj">
